+++ Example/AppDelegate+CarPlay.swift
@@ -86,8 +86,8 @@ extension AppDelegate: CarPlayManagerDelegate {
                                        simulating: desiredSimulationMode)
     }
     
-    func carPlayManager(_ carPlayManager: CarPlayManager, didBeginNavigationWith service: NavigationService) {
-        currentAppRootViewController?.beginNavigationWithCarPlay(navigationService: service)
+    func carPlayManager(_ carPlayManager: CarPlayManager, didPresent navigationViewController: CarPlayNavigationViewController) {
+        currentAppRootViewController?.beginNavigationWithCarPlay(navigationService: navigationViewController.navigationService)
         carPlayManager.carPlayNavigationViewController?.compassView.isHidden = false
         
         // Render part of the route that has been traversed with full transparency, to give the illusion of a disappearing route.
@@ -216,13 +216,13 @@ extension AppDelegate: CarPlayManagerDelegate {
                 return nil
             }
             
-            var mapButtons = [
+            let mapButtons = [
                 carPlayMapViewController.recenterButton,
+                carPlayMapViewController.panningInterfaceDisplayButton(for: mapTemplate),
                 carPlayMapViewController.zoomInButton,
                 carPlayMapViewController.zoomOutButton
             ]
             
-            mapButtons.insert(carPlayMapViewController.panningInterfaceDisplayButton(for: mapTemplate), at: 1)
             return mapButtons
         case .previewing, .navigating, .panningInBrowsingMode:
             return nil