+++ Tests/MapboxCoreNavigationTests/OptionsTests.swift
@@ -25,4 +25,20 @@ class OptionsTests: TestCase {
         XCTAssertTrue(options.includesVisualInstructions)
         XCTAssertTrue(options.includesSpokenInstructions)
     }
+    
+    func testDefaultAttributeOptions() {
+        XCTAssertEqual(NavigationRouteOptions(coordinates: coordinates).attributeOptions,
+                       [.numericCongestionLevel, .expectedTravelTime, .maximumSpeedLimit])
+        XCTAssertEqual(NavigationRouteOptions(coordinates: coordinates, profileIdentifier: .automobile).attributeOptions,
+                       [.numericCongestionLevel, .expectedTravelTime, .maximumSpeedLimit])
+        XCTAssertEqual(NavigationRouteOptions(coordinates: coordinates, profileIdentifier: .automobileAvoidingTraffic).attributeOptions,
+                       [.numericCongestionLevel, .expectedTravelTime, .maximumSpeedLimit])
+        // https://github.com/mapbox/mapbox-navigation-ios/issues/3495
+        XCTAssertEqual(NavigationRouteOptions(coordinates: coordinates, profileIdentifier: .cycling).attributeOptions,
+                       [.congestionLevel, .expectedTravelTime, .maximumSpeedLimit])
+        XCTAssertEqual(NavigationRouteOptions(coordinates: coordinates, profileIdentifier: .walking).attributeOptions,
+                       [.numericCongestionLevel, .expectedTravelTime, .maximumSpeedLimit])
+        XCTAssertEqual(NavigationRouteOptions(coordinates: coordinates, profileIdentifier: .init(rawValue: "mapbox/unicycling")).attributeOptions,
+                       [.numericCongestionLevel, .expectedTravelTime, .maximumSpeedLimit])
+    }
 }